refactor: use spring-util
