+++ build.gradle
@@ -111,6 +111,7 @@ subprojects {
             // My Utilities
             dependency 'moe.pine:heroku-addons:0.3.3'
             dependency 'moe.pine:spring-cache-interceptors:0.2.0'
+            dependency 'moe.pine:spring-util:0.1.1'
             dependency 'moe.pine:reactor-interruptible:0.1.1'
 
             // Test
