+++ lib-bookmeter/build.gradle
@@ -1,6 +1,4 @@
 dependencies {
-    implementation project(':lib-spring-utils')
-
     // Spring Framework
     implementation 'org.springframework.data:spring-data-commons'
     implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -16,6 +14,7 @@ dependencies {
     // Utilities
     implementation 'com.google.guava:guava'
     implementation 'moe.pine:reactor-interruptible'
+    implementation 'moe.pine:spring-util'
 
     // HTML parser
     implementation 'org.jsoup:jsoup'
