--- lib-spring-utils/build.gradle
@@ -1,5 +1,0 @@
-dependencies {
-    // Spring Framework
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-}
